Make generic version of JS for checkbox controls
Makes one version of the JS for the version of the
sticky nav that renders controls for a list of
checkboxes which change depending on the
selection.

This code is made generic by moving out specific
data onto data attributes in the HTML.

Variations of this generic version are created by
altering the instances it creates via factory
functions.
