+++ gulpfile.js
@@ -121,6 +121,7 @@ const javascripts = () => {
     paths.src + 'javascripts/radios-with-images.js',
     paths.src + 'javascripts/previewPane.js',
     paths.src + 'javascripts/colourPreview.js',
+    paths.src + 'javascripts/liveCheckboxControls.js',
     paths.src + 'javascripts/templateFolderForm.js',
     paths.src + 'javascripts/addBrandingOptionsForm.js',
     paths.src + 'javascripts/setAuthTypeForm.js',
